requirements.txt
google-adk
distro
langchain
langchain-community
litellm
google-cloud-aiplatform
markdown
scipy
scikit-learn
pandas
numpy
lightgbm
langchain-google-vertexai
neo4j
fastapi
uvicorn
python-dotenv
pydantic==2.11.7
unstructured
joblib
yfinance
praw
textblob
vaderSentiment